+++ docs/docs/.vitepress/theme/src/list.ts
@@ -1,4 +1,6 @@
-// 底部链接
+/**
+ * 底部页脚链接
+ */
 export const bottomList = [
   {
     title: '社区',
@@ -10,7 +12,7 @@ export const bottomList = [
         link: 'https://github.com/FightingDesign/fighting-design'
       },
       { text: 'Npm 组织', link: 'https://www.npmjs.com/org/fighting-design' },
-      { text: 'Npm 仓库', link: 'https://www.npmjs.com/package/fighting-design' },
+      { text: 'Npm 仓库', link: 'https://www.npmjs.com/package/fighting-design' }
     ]
   },
   {