+++ packages/fighting-design/__test__/table.spec.ts
@@ -1,0 +1,10 @@
+import { mount } from '@vue/test-utils'
+import { describe, expect, test } from 'vitest'
+import { FTable } from '../index'
+
+describe('Table', () => {
+  test('class', () => {
+    const wrapper = mount(FTable)
+    expect(wrapper.classes()).toContain('f-table')
+  })
+})