+++ packages/fighting-design/_utils/change-color/index.ts
@@ -11,7 +11,7 @@ export class ChangeColor implements ChangeColorInterface {
   }
   /**
    * 将 hex 色号转换为 rgb
-   * @returns
+   * @returns { String[] | void }
    */
   hexToRgb = (): string[] | void => {
     const r = /^\#?[0-9A-Fa-f]{6}$/
@@ -28,7 +28,7 @@ export class ChangeColor implements ChangeColorInterface {
   /**
    * 将 rgb 色号转换为 hex
    * @param rgb rgb 色号
-   * @returns
+   * @returns { String }
    */
   rgbToHex = (...rgb: string[]): string => {
     const hex: string[] = [...rgb]
@@ -42,7 +42,7 @@ export class ChangeColor implements ChangeColorInterface {
   /**
    * 加深颜色
    * @param level 加深程度
-   * @returns
+   * @returns { String }
    */
   getDarkColor = (level: number): string => {
     const rgb: string[] = this.hexToRgb() as string[]
@@ -54,7 +54,7 @@ export class ChangeColor implements ChangeColorInterface {
   /**
    * 减淡颜色
    * @param level 减淡程度
-   * @returns
+   * @returns { String }
    */
   getLightColor = (level: number): string => {
     const rgb: string[] = this.hexToRgb() as string[]