+++ packages/fighting-design/image-preview/src/image-preview.vue
@@ -30,8 +30,17 @@
   const prop: ImagePreviewPropsType = defineProps(Props)
   const emit = defineEmits(Emits)
 
-  const { scale, rotate, smaller, bigger, onImgMousewheel, recovery } =
-    useOperationImg()
+  const {
+    scale,
+    rotate,
+    smaller,
+    bigger,
+    scrollZoom,
+    recovery,
+    rotateClockwise,
+    rotateCounterClock
+  } = useOperationImg()
+
   const isVisible: Ref<boolean> = ref<boolean>(prop.visible)
   const previewShowIndex: Ref<number> = ref<number>(
     prop.showIndex > prop.imgList.length - 1 ? 0 : prop.showIndex
@@ -43,6 +52,7 @@
     prop.close && prop.close()
   }
 
+  // 监视绑定值，如果为假，则关闭
   watch(
     (): boolean => isVisible.value,
     (newVal: boolean): void => {
@@ -105,12 +115,8 @@
       1: (): void => smaller(),
       2: (): void => bigger(),
       3: (): void => recovery(),
-      4: (): void => {
-        rotate.value += 90
-      },
-      5: (): void => {
-        rotate.value -= 90
-      }
+      4: (): void => rotateClockwise(),
+      5: (): void => rotateCounterClock()
     } as const
 
     if (optionMap[target.key as string]) {
@@ -120,7 +126,7 @@
 </script>
 
 <template>
-  <div class="f-image-preview" @mousewheel="onImgMousewheel">
+  <div class="f-image-preview" @mousewheel="scrollZoom">
     <f-popup v-model:visible="isVisible" :open="imagPreload">
       <img
         class="f-image-preview__exhibition"
@@ -162,7 +168,7 @@
         v-if="isOption"
         class="f-image-preview__option"
         round
-        @click="optionClick"
+        :click="optionClick"
       >
         <f-toolbar-item :icon="FIconZoomOutVue" :data-key="1" />
         <f-toolbar-item :icon="FIconZoomInVue" :data-key="2" />